backend/requirements.txt
# Add libraries as needed
fastapi
uvicorn
spacy
openai
azure-storage-blob
jwt
fitz